Moced indexing-info view to indexing_views because that is more logical
