+++ src/snovault/indexing_views.py
@@ -6,11 +6,13 @@ from pyramid.security import (
 )
 from pyramid.traversal import resource_path
 from pyramid.view import view_config
+from timeit import default_timer as timer
 from .resources import Item
 from .authentication import calc_principals
 from .interfaces import STORAGE
 
 def includeme(config):
+    config.add_route('indexing-info', '/indexing-info')
     config.scan(__name__)
 
 
@@ -171,3 +173,47 @@ def item_index_data(context, request):
     }
 
     return document
+
+
+@view_config(route_name='indexing-info', permission='index', request_method='GET')
+def indexing_info(request):
+    """
+    Endpoint to check some indexing-related properties of a given uuid, which
+    is provided using the `uuid=` query parameter. This route cannot be defined
+    with the context of a specific Item because that will cause the underlying
+    request to use a cached view from Elasticsearch and not properly run
+    the @@embedded view from the database.
+
+    If you do not want to calculate the embedded object, use `run=False`
+
+    Args:
+        request: current Request object
+
+    Returns:
+        dict response
+    """
+    uuid = request.params.get('uuid')
+    if not uuid:
+        return {'status': 'error', 'title': 'Error', 'message': 'ERROR! Provide a uuid to the query.'}
+
+    db_sid = request.registry[STORAGE].write.get_by_uuid(uuid).sid
+    es_sid = request.registry[STORAGE].read.get_by_uuid(uuid).sid
+    response = {'sid_db': db_sid, 'sid_es': es_sid, 'title': 'Indexing Info for %s' % uuid}
+    if asbool(request.params.get('run', True)):
+        request._indexing_view = True
+        request.datastore = 'database'
+        path = '/' + uuid + '/@@embedded'
+        start = timer()
+        embedded_view = request.invoke_view(path, index_uuid=uuid, as_user='INDEXER')
+        end = timer()
+        response['embedded_seconds'] = end - start
+        find_uuids = get_rev_linked_items(request, uuid)
+        response['uuids_rev_linked_to_this'] = list(find_uuids)
+        find_uuids.add(uuid)
+        assc_uuids = find_uuids_for_indexing(request.registry, find_uuids)
+        response['uuids_invalidated_by_this'] = list(assc_uuids)
+        response['description'] = 'Using live results for embedded view of %s. Query with run=False to skip this.' % uuid
+    else:
+        response['description'] = 'Query with run=True to calculate live information on invalidation and embedding time.'
+    response['status'] = 'success'
+    return response
